+++ r_light.c
@@ -41,7 +41,7 @@ void r_light_start(void)
 		for (x = 0;x < 32;x++)
 		{
 			dx = (x - 15.5f) * (1.0f / 16.0f);
-			a = ((1.0f / (dx * dx + dy * dy + 0.2f)) - (1.0f / (1.0f + 0.2))) * 32.0f / (1.0f / (1.0f + 0.2));
+			a = (int)(((1.0f / (dx * dx + dy * dy + 0.2f)) - (1.0f / (1.0f + 0.2))) * 32.0f / (1.0f / (1.0f + 0.2)));
 			a = bound(0, a, 255);
 			pixels[y][x][0] = a;
 			pixels[y][x][1] = a;
