+++ world_cs.c
@@ -322,7 +322,7 @@ void CSSV_LinkEdict (prvm_edict_t *ent, qboolean touch_triggers)
 
 	if (ent->fields.client->solid == SOLID_BSP)
 	{
-		int modelindex = ent->fields.client->modelindex;
+		int modelindex = (int)ent->fields.client->modelindex;
 		if (modelindex < 0 || modelindex > MAX_MODELS)
 		{
 			Con_Printf("edict %i: SOLID_BSP with invalid modelindex!\n", PRVM_NUM_FOR_EDICT(ent));
@@ -450,7 +450,7 @@ trace_t CSSV_ClipMoveToEntity(prvm_edict_t *ent, const vec3_t start, const vec3_
 
 	if ((int) ent->fields.client->solid == SOLID_BSP || movetype == MOVE_HITMODEL)
 	{
-		unsigned int modelindex = ent->fields.client->modelindex;
+		unsigned int modelindex = (unsigned int)ent->fields.client->modelindex;
 		// if the modelindex is 0, it shouldn't be SOLID_BSP!
 		if (modelindex == 0)
 		{
